+++ types/components/base/input/search.d.ts
@@ -1,5 +1,6 @@
 /// <reference types="react" />
 export interface CustomSearchProps {
+    value?: string;
     onChange?(value: string): void;
     onKeyDown?(e: React.KeyboardEvent<HTMLButtonElement>): void;
     placeholder: string;