+++ types/components/hooks/useGetImage.d.ts
@@ -1,0 +1,4 @@
+/**
+ * Returns the image that is currently loaded in the editor.
+ */
+export declare const useGetImage: () => string | HTMLImageElement | null;